chore: code restructuring test 1
